Entity property values that are NUMBERs should be cast as a Long and not Integer

+++ interaction-core/src/main/java/com/temenos/interaction/core/entity/EntityMetadata.java
@@ -157,7 +157,7 @@ public class EntityMetadata  {
 		}
 		else if ( isPropertyNumber( propertyName ) )
 		{
-			value = Integer.toString( (Integer) property.getValue() );
+			value = Long.toString( (Long) property.getValue() );
 		}
 		
 		return value;
